2_year/2_semester/bd/scripts_python/requirements.txt
psycopg2-binary
mysql-connector-python
python-dotenv
faker
faker-vehicle